Moved `#applicable_degree` to concern on the ApplicationForm
